+++ R/estimate_R0.R
@@ -2,7 +2,7 @@
 #' Estimate the time varying R0 - using EpiEstim
 #'
 #' @param cases A dataframe containing a list of local cases with the following variables: `date`, `cases`, and `import_status`
-#' @param serial_intervals A matrix with columns representing samples and rows representing the probability of the serial intervel being on
+#' @param generation_times A matrix with columns representing samples and rows representing the probability of the generation timebeing on
 #' that day.
 #' @param rt_prior A list defining the reproduction number prior containing the mean (`mean_prior`) and standard deviation (`std_prior`)
 #' @param windows Numeric vector, windows over which to estimate time-varying R. The best performing window will be 
@@ -23,7 +23,7 @@
 #'
 #' ## Nowcast Rts                  
 #' estimates <- estimate_R0(cases = EpiSoon::example_obs_cases, 
-#'                          serial_intervals = as.matrix(EpiNow::covid_serial_intervals[,1]), 
+#'                          generation_times = as.matrix(EpiNow::covid_generation_times[,1]), 
 #'                          rt_prior = list(mean_prior = 2.6, std_prior = 2),
 #'                          windows = c(1, 3, 7), rt_samples = 10, si_samples = 2,
 #'                          min_est_date =  as.Date("2020-02-18"))
@@ -33,7 +33,7 @@
 #'   
 #'## Nowcast Rts, forecast Rts and the forecast cases
 #' estimates <- estimate_R0(cases = EpiSoon::example_obs_cases, 
-#'                          serial_intervals = as.matrix(EpiNow::covid_serial_intervals[,1]), 
+#'                          generation_times = as.matrix(EpiNow::covid_generation_times[,1]), 
 #'                          rt_prior = list(mean_prior = 2.6, std_prior = 2),
 #'                          windows = c(1, 3, 7), rt_samples = 10, si_samples = 2,
 #'                          min_est_date =  as.Date("2020-02-18"),
@@ -47,7 +47,7 @@
 #' 
 #' ## Case forecasts
 #' estimates$cases
-estimate_R0 <- function(cases = NULL, serial_intervals = NULL,
+estimate_R0 <- function(cases = NULL, generation_time = NULL,
                         rt_prior = NULL, windows = NULL, 
                         si_samples = 100, rt_samples = 100,
                         min_est_date = NULL, forecast_model = NULL, 
@@ -101,14 +101,14 @@ estimate_R0 <- function(cases = NULL, serial_intervals = NULL,
   
   
   ## Sample serial intervals
-  serial_intervals_index <- sample(1:ncol(serial_intervals),
+  generation_times_index <- sample(1:ncol(generation_times),
                              si_samples,
-                             replace = ncol(serial_intervals) < si_samples)
+                             replace = ncol(generation_times) < si_samples)
 
    
   ### Estimate R across serial interval samples
   ### Forecast ahead if given a model and horizon for each sample
-  estimates <- purrr::map(serial_intervals_index, function(index) {
+  estimates <- purrr::map(generation_times_index, function(index) {
     
     ### Estimate and score R over multiple windows
     est_r <- purrr::map(windows, 
@@ -124,7 +124,7 @@ estimate_R0 <- function(cases = NULL, serial_intervals = NULL,
                           R <- suppressWarnings(
                             EpiEstim::estimate_R(incid,
                                                  method = "si_from_sample",
-                                                 si_sample = serial_intervals[, index],
+                                                 si_sample = generation_times[, index],
                                                  config = do.call(EpiEstim::make_config,
                                                                   c(rt_prior, 
                                                                     list(t_start = window_start,
@@ -152,7 +152,7 @@ estimate_R0 <- function(cases = NULL, serial_intervals = NULL,
                               ~ EpiSoon::predict_current_cases(
                                 rts = .[,.(date, rt = sample_R)], 
                                 cases = summed_cases,
-                                serial_interval = serial_intervals[, index]
+                                serial_interval = generation_times[, index]
                               ))
                           
                           preds <- data.table::rbindlist(preds, idcol = "sample")
@@ -221,7 +221,7 @@ estimate_R0 <- function(cases = NULL, serial_intervals = NULL,
               cases = summed_cases,
               fit_samples = rt_forecasts,
               rdist = rpois,
-              serial_interval = serial_intervals[, index]
+              serial_interval = generation_times[, index]
             ) 
           )[,rt_type := "forecast"]