+++ NEWS.md
@@ -7,6 +7,7 @@
 * Added mapping from onset cases to infection with upscaling to account for additional truncation.
 * Added minimal example to all functions
 * Added an approximate mapping for reporting delay that has an adaptive case threshold.
+* Switched to using a generation time estimate rather than a serial interval (see `data-raw/ganyani-generation-time.rs)
 
 # EpiNow 0.1.0